backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
bcrypt==4.2.1
openpyxl==3.1.5
reportlab==4.2.5
pypdf==5.2.0
psycopg[binary]==3.3.3
python-dotenv==1.0.1
gunicorn==23.0.0
